+++ changelog.md
@@ -4,7 +4,7 @@
 - Add more tests and more error handling to better survive wrong instructions
 like buying too many shares.
 - Automatic private company income when OR changes.
-- Different halfdividend roundings: round for company, calculate exactly round halves for bank.
+- Company credits.
 
 ## Unreleased (available in `master`)
 - An external cash display support is available! If the environmental variable
@@ -22,6 +22,12 @@ tests.
 - Added: If you want to specify a currency other than the default dollar, you
 can use any single-character currency symbol when setting the bank size:
 `banksize £2000` would set pound as the currency.
+- Added: `rounding up` changes the half-dividend company rounding to round in
+favour of the company (as in 1858, 1869, 1870, 18??, 18EZ, 18GM and 18NK),
+`rounding 1837` for 1837 style rounding (count exact sum, round dividend
+payments down).
+- Added: `companycash` switches the game to use company credits, where company
+cash is not part of the bank.
 
 ## 2.0.0 – 2019-07-08
 - Removed: `give` and `take` have been replaced by `cash`.